+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- **DatePicker** can be used with forwarded ref now.
+
 ## [8.23.1] - 2019-03-14
 
 ### Fixed